backend/requirements.txt
django
django-ninja
python-dotenv
Pillow